adds read rfid feature
Adds functions to initialize rfid module, detect if a card is present and read the UID of the card as String. These functions should be extracted to another files for better modularity.

This code compiles but it has not been tested

This file is heavily based on MFRC522 library examples.